Added update feature that makes working with external sprites easier
The path to the texture of the imported asset is saved in the
RagePixelCell. Options to update the selected frame, sprite, or all
sprites are implemented. Does not maintain a connection with imported
sprite sheets.

If textures are moved just replace the existing frame and the
connection will be setup again. The name of the file is listed beside
the frame index.

Bug Fix: Imported sprites were scaling to the nearest power of 2.